// src/app/common/types/thing.type.ts
export interface ThingType {
  uuid: string;
  associateId: string;
  name?: string;
  type: string;
  url?: string;
  gender?: string;
  counter?: number;
  thumbnailUrl?: string;
  color?: string;
  comment?: string;
  firstColour?: string;
  secondColour?: string;
  thirdColour?: string;
  template?: boolean;
  action?: string
}